+++ crates/evm-coder/src/lib.rs
@@ -16,6 +16,7 @@
 
 #![doc = include_str!("../README.md")]
 #![deny(missing_docs)]
+#![macro_use]
 #![cfg_attr(not(feature = "std"), no_std)]
 #[cfg(not(feature = "std"))]
 extern crate alloc;
@@ -26,6 +27,8 @@ pub mod abi;
 pub use events::{ToLog, ToTopic};
 use execution::DispatchInfo;
 pub mod execution;
+#[macro_use]
+pub mod custom_signature;
 
 /// Derives call enum implementing [`crate::Callable`], [`crate::Weighted`]
 /// and [`crate::Call`] from impl block.
@@ -90,6 +93,7 @@ pub use evm_coder_procedural::solidity_interface;
 pub use evm_coder_procedural::solidity;
 /// See [`solidity_interface`]
 pub use evm_coder_procedural::weight;
+pub use sha3_const;
 
 /// Derives [`ToLog`] for enum
 ///
@@ -104,6 +108,7 @@ pub use evm_coder_procedural::ToLog;
 #[doc(hidden)]
 pub mod events;
 #[doc(hidden)]
+#[cfg(feature = "stubgen")]
 pub mod solidity;
 
 /// Solidity type definitions (aliases from solidity name to rust type)
@@ -114,28 +119,55 @@ pub mod types {
 
 	#[cfg(not(feature = "std"))]
 	use alloc::{vec::Vec};
+	use pallet_evm::account::CrossAccountId;
 	use primitive_types::{U256, H160, H256};
+	use core::str::from_utf8;
 
-	pub type address = H160;
+	use crate::custom_signature::SignatureUnit;
 
-	pub type uint8 = u8;
-	pub type uint16 = u16;
-	pub type uint32 = u32;
-	pub type uint64 = u64;
-	pub type uint128 = u128;
-	pub type uint256 = U256;
+	pub trait Signature {
+		const SIGNATURE: SignatureUnit;
 
-	pub type bytes4 = [u8; 4];
+		fn as_str() -> &'static str {
+			from_utf8(&Self::SIGNATURE.data[..Self::SIGNATURE.len]).expect("bad utf-8")
+		}
+	}
 
-	pub type topic = H256;
+	impl Signature for bool {
+		const SIGNATURE: SignatureUnit = make_signature!(new fixed("bool"));
+	}
+
+	macro_rules! define_simple_type {
+		(type $ident:ident = $ty:ty) => {
+			pub type $ident = $ty;
+			impl Signature for $ty {
+				const SIGNATURE: SignatureUnit = make_signature!(new fixed(stringify!($ident)));
+			}
+		};
+	}
+
+	define_simple_type!(type address = H160);
+
+	define_simple_type!(type uint8 = u8);
+	define_simple_type!(type uint16 = u16);
+	define_simple_type!(type uint32 = u32);
+	define_simple_type!(type uint64 = u64);
+	define_simple_type!(type uint128 = u128);
+	define_simple_type!(type uint256 = U256);
+	define_simple_type!(type bytes4 = [u8; 4]);
+
+	define_simple_type!(type topic = H256);
 
 	#[cfg(not(feature = "std"))]
-	pub type string = ::alloc::string::String;
+	define_simple_type!(type string = ::alloc::string::String);
 	#[cfg(feature = "std")]
-	pub type string = ::std::string::String;
+	define_simple_type!(type string = ::std::string::String);
 
 	#[derive(Default, Debug)]
 	pub struct bytes(pub Vec<u8>);
+	impl Signature for bytes {
+		const SIGNATURE: SignatureUnit = make_signature!(new fixed("bytes"));
+	}
 
 	/// Solidity doesn't have `void` type, however we have special implementation
 	/// for empty tuple return type
@@ -166,6 +198,7 @@ pub mod types {
 		}
 	}
 
+	#[allow(clippy::from_over_into)]
 	impl Into<Vec<u8>> for bytes {
 		fn into(self) -> Vec<u8> {
 			self.0
@@ -182,6 +215,76 @@ pub mod types {
 		pub fn is_empty(&self) -> bool {
 			self.len() == 0
 		}
+	}
+
+	#[derive(Debug, Default)]
+	pub struct EthCrossAccount {
+		pub(crate) eth: address,
+		pub(crate) sub: uint256,
+	}
+
+	impl EthCrossAccount {
+		pub fn from_sub_cross_account<T>(cross_account_id: &T::CrossAccountId) -> Self
+		where
+			T: pallet_evm::account::Config,
+			T::AccountId: AsRef<[u8; 32]>,
+		{
+			if cross_account_id.is_canonical_substrate() {
+				Self {
+					eth: Default::default(),
+					sub: convert_cross_account_to_uint256::<T>(cross_account_id),
+				}
+			} else {
+				Self {
+					eth: *cross_account_id.as_eth(),
+					sub: Default::default(),
+				}
+			}
+		}
+
+		pub fn into_sub_cross_account<T>(&self) -> crate::execution::Result<T::CrossAccountId>
+		where
+			T: pallet_evm::account::Config,
+			T::AccountId: From<[u8; 32]>,
+		{
+			if self.eth == Default::default() && self.sub == Default::default() {
+				Err("All fields of cross account is zeroed".into())
+			} else if self.eth == Default::default() {
+				Ok(convert_uint256_to_cross_account::<T>(self.sub))
+			} else if self.sub == Default::default() {
+				Ok(T::CrossAccountId::from_eth(self.eth))
+			} else {
+				Err("All fields of cross account is non zeroed".into())
+			}
+		}
+	}
+
+	impl Signature for EthCrossAccount {
+		const SIGNATURE: SignatureUnit = make_signature!(new fixed("(address,uint256)"));
+	}
+
+	/// Convert `CrossAccountId` to `uint256`.
+	pub fn convert_cross_account_to_uint256<T: pallet_evm::account::Config>(
+		from: &T::CrossAccountId,
+	) -> uint256
+	where
+		T::AccountId: AsRef<[u8; 32]>,
+	{
+		let slice = from.as_sub().as_ref();
+		uint256::from_big_endian(slice)
+	}
+
+	/// Convert `uint256` to `CrossAccountId`.
+	pub fn convert_uint256_to_cross_account<T: pallet_evm::account::Config>(
+		from: uint256,
+	) -> T::CrossAccountId
+	where
+		T::AccountId: From<[u8; 32]>,
+	{
+		let mut new_admin_arr = [0_u8; 32];
+		from.to_big_endian(&mut new_admin_arr);
+		let account_id = T::AccountId::from(new_admin_arr);
+		T::CrossAccountId::from_sub(account_id)
 	}
 }
 
@@ -239,7 +342,7 @@ impl Call for ERC165Call {
 			return Ok(None);
 		}
 		Ok(Some(Self::SupportsInterface {
-			interface_id: input.abi_read()?,
+			interface_id: types::bytes4::abi_read(input)?,
 		}))
 	}
 }
@@ -252,6 +355,7 @@ impl Call for ERC165Call {
 #[macro_export]
 macro_rules! generate_stubgen {
 	($name:ident, $decl:ty, $is_impl:literal) => {
+		#[cfg(feature = "stubgen")]
 		#[test]
 		#[ignore]
 		fn $name() {